Move code out of org.langmeta.internal.semanticdb._
