+++ langmeta/langmeta/shared/src/main/scala/org/langmeta/internal/semanticdb/package.scala
@@ -1,0 +1,11 @@
+package org.langmeta.internal
+
+import scala.meta.internal.{semanticdb3 => s}
+
+package object semanticdb {
+
+  final val DeprecationMessage =
+    "This API is no longer supported. Use scala.meta.internal.semanticdb3._ instead, but note that this API is subject to binary and source breaking changes without notice."
+
+
+}
