+++ tests/jvm/src/test/scala/scala/meta/tests/semanticdb/TargetedSuite.scala
@@ -4,6 +4,7 @@ package semanticdb
 import org.langmeta.internal.semanticdb._
 import scala.meta._
 import scala.meta.internal.semanticdb.scalac._
+import scala.meta.internal.semanticdb3._
 import scala.meta.internal.semanticdb3.SymbolInformation.{Kind => k}
 import scala.meta.internal.semanticdb3.SymbolInformation.{Property => p}
 import Compat._
